requirements.txt
flox-lib==0.19.6
cattrs
requests
linguee-api
detectlanguage